move Joysticks functions to RobotMotorsSetup
