+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Supernova11567Main.java
@@ -63,32 +63,4 @@ public class Supernova11567Main extends OpMode {
     public void stop() {
     }
 
-    public double getJoystickXValue(double angle) {
-        double newAngleX = angle;
-        double xValue;
-
-        if (Math.abs(newAngleX) > 179) {
-            newAngleX = 0;
-        } else {
-            newAngleX = Range.clip(newAngleX, -179, 179);
-        }
-
-        if (Math.abs(newAngleX) > 90) {
-            xValue = ((Math.abs(newAngleX) / 90) - 2 * ((Math.abs(newAngleX) / 90) % 1)) * (newAngleX / Math.abs(newAngleX));
-            return xValue;
-        } else {
-            xValue = newAngleX / 90;
-            return xValue;
-        }
-    }
-
-
-    public double getJoystickYValue(double angle) {
-        double newAngleY = angle;
-        double yValue;
-
-        yValue = (Math.abs(newAngleY) / 90) - 1;
-
-        return yValue;
-    }
 }
